Slightly hacky fix for yaourt and git $HOME.
Since I automatically set GIT_DIR and GIT_WORK_TREE when in
home dir for so that I can seamlessly use git I've noticed that
yaourt gets confused when cloning git repos (and I'm in $HOME).

This works around the problem by redefining yaourt as a function
unsetting those variables. This should probably be fixed in a nicer
way.